+++ src/SMAPI/Framework/ModLoading/ModResolver.cs
@@ -137,11 +137,22 @@ namespace StardewModdingAPI.Framework.ModLoading
                         }
 
                         // invalid path
-                        string assemblyPath = Path.Combine(mod.DirectoryPath, mod.Manifest.EntryDll);
-                        if (!File.Exists(assemblyPath))
+                        if (!File.Exists(Path.Combine(mod.DirectoryPath, mod.Manifest.EntryDll)))
                         {
                             mod.SetStatus(ModMetadataStatus.Failed, $"its DLL '{mod.Manifest.EntryDll}' doesn't exist.");
                             continue;
+                        }
+
+                        // invalid capitalisation
+                        string actualFilename = new DirectoryInfo(mod.DirectoryPath).GetFiles(mod.Manifest.EntryDll).FirstOrDefault()?.Name;
+                        if (actualFilename != mod.Manifest.EntryDll)
+                        {
+#if SMAPI_3_0_STRICT
+                            mod.SetStatus(ModMetadataStatus.Failed, $"its {nameof(IManifest.EntryDll)} value '{mod.Manifest.EntryDll}' doesn't match the actual file capitalisation '{actualFilename}'. The capitalisation must match for crossplatform compatibility.");
+                            continue;
+#else
+                            SCore.DeprecationManager.Warn(mod.DisplayName, $"{nameof(IManifest.EntryDll)} value with case-insensitive capitalisation", "2.11", DeprecationLevel.Info);
+#endif
                         }
                     }